requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
djangorestframework==3.14.0
docker==6.1.3
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
typing-extensions==4.8.0
urllib3==2.0.5
websocket-client==1.6.3
